backend/requirements.txt
Flask
PyYAML
Flask-Cors
paramiko
pymongo
dnspython
